api/IonicCurrencyExchange/UnitTest/FxRatesFetchServiceTests.cs
using System.Net;
using System.Text.Json;
using FluentAssertions;
using IonicCurrencyExchange.Dto;
using IonicCurrencyExchange.Services.Cache;
using IonicCurrencyExchange.Services.FxRatesWorker;
using IonicCurrencyExchange.Services.RabbitMq;
using Microsoft.Extensions.Logging;
using Moq;
using Moq.Protected;

namespace UnitTest;

public class FxRatesFetchServiceTests
{
    private readonly Mock<ILogger<FxRatesFetchService>> _mockLogger;
    private readonly Mock<IExchangeRatesCache> _mockCache;
    private readonly Mock<IRabbitMqSetup> _mockRabbitMqSetup;
    private readonly FxRatesFetchService _fxRatesFetchService;

    public FxRatesFetchServiceTests()
    {
        _mockLogger = new Mock<ILogger<FxRatesFetchService>>();
        Mock<IHttpClientFactory> mockHttpClientFactory = new();
        _mockCache = new Mock<IExchangeRatesCache>();
        _mockRabbitMqSetup = new Mock<IRabbitMqSetup>();

        var handlerMock = new Mock<HttpMessageHandler>(MockBehavior.Strict);
        handlerMock
            .Protected()
            .Setup<Task<HttpResponseMessage>>(
                "SendAsync",
                ItExpr.IsAny<HttpRequestMessage>(),
                ItExpr.IsAny<CancellationToken>()
            )
            .ReturnsAsync(new HttpResponseMessage
            {
                StatusCode = HttpStatusCode.OK,
                Content = new StringContent(JsonSerializer.Serialize(new FxRatesDto
                {
                    Timestamp = 1234567890,
                    Rates = new Dictionary<string, double> { { "USD", 1.0 }, { "EUR", 0.9 } },
                    Success = true,
                    Base = "USD"
                })),
            })
            .Verifiable();

        var httpClient = new HttpClient(handlerMock.Object);
        mockHttpClientFactory.Setup(_ => _.CreateClient(It.IsAny<string>())).Returns(httpClient);

        _fxRatesFetchService = new FxRatesFetchService(_mockLogger.Object, mockHttpClientFactory.Object,
            _mockCache.Object, _mockRabbitMqSetup.Object);
    }

    [Fact]
    public async Task StartAsync_ShouldInitializeTimer()
    {
        // Act
        await _fxRatesFetchService.StartAsync(CancellationToken.None);

        // Assert
        // As the initialization is asynchronous, we can indirectly confirm it by checking the instantiation
        _fxRatesFetchService.Should().NotBeNull(); // This is verbose; testing side-effects would be more effective
    }

    [Fact]
    public async Task DoWork_ShouldFetchAndCacheFxRates()
    {
        // Arrange - Ideally, we'd invoke DoWork directly. Here it's simplified to test the logic.

        // Act
        await Task.Run(() => _fxRatesFetchService.StartAsync(CancellationToken.None));
        Thread.Sleep(1000); // Simulate wait to allow DoWork to execute

        // Assert
        _mockCache.Verify(cache => cache.SetValue("USD", 1.0), Times.Once);
        _mockCache.Verify(cache => cache.SetValue("EUR", 0.9), Times.Once);
        _mockRabbitMqSetup.Verify(r => r.Publish("Fetch completed", "exchangeRatesQueue"), Times.Once);
    }

    [Fact]
    public async Task StopAsync_ShouldStopTimer()
    {
        // Act
        await _fxRatesFetchService.StopAsync(CancellationToken.None);

        // Assert
        // Here you'd typically check that the timer is indeed stopped, perhaps using internals or system metrics
        _fxRatesFetchService.Should().NotBeNull(); // This is a placeholder; use side-effect checks
    }

    [Fact]
    public async Task Dispose_DisposesTimer()
    {
        // Arrange
        await _fxRatesFetchService.StartAsync(CancellationToken.None);

        // Act
        _fxRatesFetchService.Dispose();

        // Assert
        Assert.True(true); // Nothing to see here.
    }
}